# lib/api-zod/src/generated/types/legacyMigrationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Neurohacking account, migration and technique API
 * OpenAPI spec version: 0.1.0
 */
import type { LegacyMigrationInputState } from './legacyMigrationInputState';

export interface LegacyMigrationInput {
  /**
     * @minLength 16
     * @maxLength 128
     */
  migrationKey: string;
  state: LegacyMigrationInputState;
}