# lib/api-zod/src/generated/types/serverMe.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Neurohacking account, migration and technique API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerMeState } from './serverMeState';
import type { ServerUser } from './serverUser';
import type { UserProfile } from './userProfile';

export interface ServerMe {
  user: ServerUser;
  /** @nullable */
  state: ServerMeState;
  profile: UserProfile;
}