# lib/api-zod/src/generated/types/serverUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Neurohacking account, migration and technique API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerUser {
  id: number;
  clerkId: string;
  /** @nullable */
  email: string | null;
}